Move LinearLeastSquares from Interpolation to LinearAlgebra
The SphericalHarmonics library is needed by CardinalInterpolator, but that
creates a circular dependency because Strahlkorper uses LinearLeastSquares.
Moving LinearLeastSquares to LinearAlgebra fixes this and also makes more sense.
