+++ tests/Unit/NumericalAlgorithms/Interpolation/Test_PredictedZeroCrossing.cpp
@@ -8,8 +8,8 @@
 #include <vector>
 
 #include "Framework/TestHelpers.hpp"
-#include "NumericalAlgorithms/Interpolation/LinearLeastSquares.hpp"
 #include "NumericalAlgorithms/Interpolation/PredictedZeroCrossing.hpp"
+#include "NumericalAlgorithms/LinearAlgebra/LinearLeastSquares.hpp"
 
 namespace {
